Fixing warnings on Windows (JIRA Issue AXIS2C-573)

+++ util/include/axutil_hash.h
@@ -48,7 +48,7 @@ extern "C"
      *         Since the hash table merely references the provided key (rather
      *         than copying it), axutil_hash_this() will return the NUL-term'd key.
      */
-#define AXIS2_HASH_KEY_STRING     (-1)
+#define AXIS2_HASH_KEY_STRING     (unsigned int)(-1)
 
     /**
      * Abstract type for hash tables.
